requirements.txt
Django == 1.9
Pillow
mysqlclient == 1.3
django-crispy-forms==1.6.0
